feat(server): server emulator
Server Emulator with Lerna, yarn and typescript setup
