+++ .gitignore
@@ -1,5 +1,6 @@
 node_modules/
-npm-debug.log
+*.log
 .idea/
 .vscode/
 .DS_Store
+dist/
