--- .npmrc
@@ -1,3 +1,0 @@
-registry=https://registry.npmjs.org/
-strict-ssl=false
-always-auth=false
