+++ .yarnrc
@@ -1,0 +1,3 @@
+registry "https://registry.npmjs.org/"
+strict-ssl false
+always-auth false
