+++ lerna.json
@@ -1,7 +1,14 @@
 {
-  "lerna": "2.5.1",
-  "packages": [
-    "packages/*"
-  ],
-  "version": "0.0.0"
+    "lerna": "2.5.1",
+    "packages": [
+        "packages/*"
+    ],
+    "version": "1.0.0",
+    "npmClient": "yarn",
+    "useWorkspaces": true,
+    "command": {
+        "publish": {
+            "allowBranch": "master"
+        }
+    }
 }
