+++ packages/ufa-cli/tsconfig.build.json
@@ -1,0 +1,7 @@
+{
+    "extends": "../../tsconfig.base.json",
+    "compilerOptions": {
+      "rootDir": "src",
+      "outDir": "dist"
+    }
+}
