--- packages/ufa-documentation-manager/.npmignore
@@ -1,1 +1,0 @@
-*.log
