--- packages/ufa-documentation-manager/README.md
@@ -1,2 +1,0 @@
-
-# TBA
