--- packages/ufa-package-manager/.npmignore
@@ -1,1 +1,0 @@
-*.log
