--- packages/ufa-package-manager/README.md
@@ -1,2 +1,0 @@
-
-# TBA
