--- packages/ufa-package-manager/package.json
@@ -1,19 +1,0 @@
-{
-    "name": "@ufa/ufa-package-manager",
-    "version": "1.0.0",
-    "description": "Package and Dependencies Manager",
-    "license": "MIT",
-    "repository": "https://github.com/universal-frontend-architecture/universal-frontend-architecture/tree/master/packages/ufa-package-manager",
-    "main": "./index.js",
-    "bin": {
-        "ufa-server": "./script.js"
-    },
-    "scripts": {
-        "test": "echo \"Error: no test specified\" && exit 1"
-    },
-    "dependencies": {},
-    "devDependencies": {},
-    "publishConfig": {
-        "access": "public"
-    }
-}
