--- packages/ufa-release-manager/README.md
@@ -1,2 +1,0 @@
-
-# TBA
