+++ tsconfig.json
@@ -1,0 +1,10 @@
+{
+    "extends": "./tsconfig.base.json",
+    "compilerOptions": {
+      "baseUrl": "../packages",
+      "paths": {
+        "@ufa/*": ["./*/src"]
+      }
+    }
+}
+
